* Replaced the old logger by the new one provided by Yuni  * It behaves exactly the same than Yuni::String when using the operator '<<'  * The behavior of the new logging mecanism is not strictly equivalent to the old one. The log file is not truncated at the startup (not sure that it is the duty of `Yuni::Logs::Logger`)  Example :  \code  logs.notice() << "Hello world !";  logs.info() << "A more complex example : " << 42;  \endcode

+++ trunk/src/ta3d/src/gfx/gfx.cpp
@@ -740,7 +740,7 @@ namespace TA3D
                         glTexImage2D(GL_TEXTURE_2D, 0, texture_format, bmp->w, bmp->h, 0, GL_RGBA, GL_UNSIGNED_BYTE, bmp->pixels);
 					break;
 				default:
-					LOG_DEBUG("SDL_Surface format not supported by texture loader: " << bmp->format->BitsPerPixel << " bpp" );
+					LOG_DEBUG("SDL_Surface format not supported by texture loader: " << (int) bmp->format->BitsPerPixel << " bpp" );
 			}
 
 			if (g_useGenMipMaps && glGenerateMipmapEXT && build_mipmaps)
@@ -769,7 +769,7 @@ namespace TA3D
                         glTexImage2D(GL_TEXTURE_2D, level, texture_format, w, h, 0, GL_RGBA, GL_UNSIGNED_BYTE, tmp->pixels);
 						break;
 					default:
-						LOG_DEBUG("SDL_Surface format not supported by texture loader: " << tmp->format->BitsPerPixel << " bpp" );
+						LOG_DEBUG("SDL_Surface format not supported by texture loader: " << (int) tmp->format->BitsPerPixel << " bpp" );
 				}
 				SDL_FreeSurface(tmp);
 				++level;
